Move base analytic unit model class to a separate directory
