+++ src/panel/graph_panel/graph_tooltip.ts
@@ -1,4 +1,4 @@
-import { AnalyticSegmentsSearcher } from './models/analytic_unit';
+import { AnalyticSegmentsSearcher } from './models/analytic_units/analytic_unit';
 
 
 export class GraphTooltip {
