+++ src/panel/graph_panel/services/analytic_service.ts
@@ -2,7 +2,7 @@ import { SegmentId } from '../models/segment';
 import { MetricExpanded } from '../models/metric';
 import { DatasourceRequest } from '../models/datasource';
 import { SegmentsSet } from '../models/segment_set';
-import { AnalyticUnitId, AnalyticUnit, AnalyticSegment } from '../models/analytic_unit';
+import { AnalyticUnitId, AnalyticUnit, AnalyticSegment } from '../models/analytic_units/analytic_unit';
 import { HasticServerInfo, HasticServerInfoUnknown } from '../models/hastic_server_info';
 import { Threshold } from '../models/threshold';
 import { DetectionSpan } from '../models/detection';
